Apply Avalonia PR review feedback: decimal precision, stale doc ref, macOS release body, README structure (#40)
Addresses four review comments on the Avalonia cross-platform desktop
PR.

## Changes

- **`HybridCoordinatorWindow.axaml.cs`**
- Remove reference to non-existent `docs/desktop-bridge-contract.md` in
class summary
- Use `u.GetDecimal()` instead of `(decimal)u.GetDouble()` for
`unitPrice` parsing — avoids IEEE 754 double→decimal precision loss for
financial values (consistent with WebView2App)

```csharp
// Before
? (decimal)u.GetDouble()
// After
? u.GetDecimal()
```

- **`release-desktop-crossplatform.yml`** — Add missing `body:` block to
the macOS versioned-tag release step; was entirely absent unlike the
Linux step, leaving macOS releases with no install/usage instructions

- **`README.md`** — Add `DesktopHost.Avalonia/` directory tree to the
Project Structure section alongside the existing `WebView2App/` entry

`LocalAssetServer.cs` path traversal protection was already in place
from a prior commit; no change needed.

+++ DesktopHost.Avalonia/HybridCoordinatorWindow.axaml.cs
@@ -10,7 +10,7 @@ namespace DragDropAvaloniaDemo;
 /// <summary>
 /// Hybrid coordinator window: source is the web-based Available Items table
 /// (HTML/JS inside an embedded WebView), target is a native Avalonia DataGrid.
-/// Implements the Desktop Bridge Contract (see docs/desktop-bridge-contract.md).
+/// Implements the Desktop Bridge Contract between the web source and native target.
 /// </summary>
 public partial class HybridCoordinatorWindow : Window
 {
@@ -127,7 +127,7 @@ public partial class HybridCoordinatorWindow : Window
         decimal unitPrice = 0m;
         if (root.TryGetProperty("unitPrice", out var u))
             unitPrice = u.ValueKind == JsonValueKind.Number
-                ? (decimal)u.GetDouble()
+                ? u.GetDecimal()
                 : decimal.TryParse(u.GetString(), System.Globalization.NumberStyles.Any,
                     System.Globalization.CultureInfo.InvariantCulture, out var ud) ? ud : 0m;
